Restructure project with cosmocentric physics domains
Organized codebase into core physics domains:
- energy (electromagnetic, thermodynamics)
- forces (fundamental interactions)
- matter (physical substances, fluids)
- information (patterns, fractals)
- spacetime (dimensional framework)

Moved utility components to systems folder.
Added comprehensive descriptions to all crates.
Improved dependency relationships between components.

This should make the architecture ideal and highly scalable, might still benefits from minor enhancements later on.
This restructuring follows a cosmocentric approach that mirrors fundamental physics domains, making the architecture more intuitive and aligned with natural systems.
