+++ crates/energy/src/lib.rs
@@ -1,0 +1,2 @@
+pub use electromagnetic;
+pub use thermodynamics;
