+++ crates/forces/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "forces"
+version = "0.1.0"
+edition = "2021"
+description = "Fundamental interaction mechanisms that create relationships between entities and drive dynamic behaviors across scales."
+
+[dependencies]
