+++ crates/information/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "information"
+version = "0.1.0"
+edition = "2021"
+description = "Systems handling pattern emergence, complexity formation, and the organization principles that enable order from chaos."
+
+[dependencies]
+fractals = { path = "./fractals" }
