+++ crates/information/fractals/src/fractals.json
