+++ crates/information/fractals/src/fractals.rs
