+++ crates/information/fractals/src/generator.rs
