+++ crates/information/fractals/src/grammar.rs
