+++ crates/information/fractals/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod generator;
 pub mod grammar;
 pub mod interpreter;
-pub mod lsystem;
+pub mod fractals;
 pub mod renderer;
 pub mod data_loader;
