+++ crates/matter/fluids/Cargo.toml
@@ -2,6 +2,7 @@
 name = "fluids"
 version = "0.1.0"
 edition = "2021"
+description = "Modeling of continuous substance flow and behavior under various physical conditions."
 
 [dependencies]
 #Not sure that we will use this
