+++ crates/matter/src/lib.rs
@@ -1,0 +1,1 @@
+pub use fluids;
