+++ crates/systems/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "systems"
+version = "0.1.0"
+edition = "2021"
+description = "Core simulation infrastructure components that manage and support the virtual universe rather than represent parts of it."
+
+[dependencies]
+save_system = { path = "./save_system" }
