+++ examples/basic_save.rs
@@ -1,5 +1,5 @@
-use save_system::save_system::{save, load};
-use save_system::versioning::SAVE_VERSION;
+use systems::save_system::save_system::{save, load};
+use systems::save_system::versioning::SAVE_VERSION;
 use serde::{Serialize, Deserialize};
 
 #[derive(Serialize, Deserialize, Debug)]
